Extension runs automatically whenever file, folder is changed

+++ src/extension.ts
@@ -1,31 +1,48 @@
-import * as vscode from 'vscode'
+import * as vscode from 'vscode';
 import * as fs from 'fs'
 import * as path from 'path'
 import ignore, { Ignore } from 'ignore'
+const onWorkspaceLoad = async () => {
+  vscode.window.showInformationMessage('Workspace Loaded!');
+};
 
-export function activate(context: vscode.ExtensionContext) {
-  let disposable = vscode.commands.registerCommand('codesnapshot.helloWorld', () => {
-    const projectDir = vscode.workspace.rootPath?.toString() ?? ''
-    const outputFile = path.join(projectDir, 'code_context.txt')
+const onFileSystemChange = async (uri: vscode.Uri) => {
+  const filePath = uri.fsPath;
+  const projectDir = vscode.workspace.rootPath?.toString() ?? ''
+  const outputFile = path.join(projectDir, 'code_context.txt')
 
-    fs.existsSync(outputFile) && fs.unlinkSync(outputFile)
+  // Check if the changed file is code_context.txt and return early if true
+  if (filePath === outputFile) {
+   return;
+  }
 
-    const ig = loadGitignore(projectDir)
+  vscode.window.showInformationMessage(`File Changed: ${filePath}`);
 
-    const rootEntries = fs.readdirSync(projectDir, { withFileTypes: true })
-    rootEntries.forEach((entry) => {
-      if (entry.isDirectory()) {
-        const dirPath = path.join(projectDir, entry.name)
-        readFiles(dirPath, outputFile, projectDir, ig)
-      }
-    })
+  fs.existsSync(outputFile) && fs.unlinkSync(outputFile)
 
-    vscode.window.showInformationMessage('Script executed correctly!')
+  const ig = loadGitignore(projectDir)
+
+  const rootEntries = fs.readdirSync(projectDir, { withFileTypes: true })
+  rootEntries.forEach((entry) => {
+    if (entry.isDirectory()) {
+      const dirPath = path.join(projectDir, entry.name)
+      readFiles(dirPath, outputFile, projectDir, ig)
+    }
   })
 
-  context.subscriptions.push(disposable)
-}
+  vscode.window.showInformationMessage('Script executed correctly!')
+};
 
+export async function activate(context: vscode.ExtensionContext) {
+  await onWorkspaceLoad();
+
+  const watcher = vscode.workspace.createFileSystemWatcher('**'); // Watch all files recursively
+  watcher.onDidChange(onFileSystemChange)
+  watcher.onDidCreate(onFileSystemChange)
+  watcher.onDidDelete(onFileSystemChange)
+
+  context.subscriptions.push(watcher);
+}
 function loadGitignore(projectDir: string) {
   const gitignorePath = path.join(projectDir, '.gitignore')
   const ig = ignore()
@@ -64,5 +81,5 @@ function appendFileIfNotIgnored(fullPath: string, outputFile: string, projectDir
     fs.appendFileSync(outputFile, '\n')
   }
 }
-
 export function deactivate() {}
+
